chore: update README
