+++ README.md
@@ -1,10 +1,22 @@
 # CODESPACERS BACKEND
 
 ## STACK
+
 Express
 Typescript
 Couchbase
 
 ## Contributing
-1. Use PNPM
 
+1. Use pnpm
+2. scripts are in package.json
+3. run pnpm install
+4. pnpm dev:start to run the app in dev mode
+5. pnpm build to build the app
+6. pnpm start to run the app, you might as well install pm2 also
+
+## LIVE LINK
+
+[backend](https://team-favourz-backend.onrender.com/api/v1/health)
+
+Docs are on /api-docs
