+++ src/controllers/subscription.controller.ts
@@ -1,0 +1,1 @@
+import type { Request, Response } from "express";
