+++ src/routes/user.routes.ts
@@ -1,15 +1,16 @@
-
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import { Router } from "express";
-import { registerUser, userLogIn, logout, createFolder, displayFolderDetails} from "../controllers/index";
+import { registerUser, userLogIn, logout } from "@/controllers/auth.controller";
+import { createFolder, displayFolderDetails } from "../controllers/index";
 import { verifyTokenFromCookie } from "../middlewares/token_verifier";
 const router = Router();
 
 router.post("/register", registerUser);
 router.post("/login", userLogIn);
 router.post("/logout", verifyTokenFromCookie, logout);
-router.post("/createfolder", createFolder)
-router.get("/displayfolderdetails", displayFolderDetails)
+
+router.post("/createfolder", createFolder);
+router.get("/displayfolderdetails", displayFolderDetails);
 
 export default router;
